Touchup function names
In general it is bad practice to put "return" in the name of a function
since it should be clear that most functions return something from
their signature.  Additionally, locking in varaible names like
"col_j" makes it difficult when your variable is not "j" such as
A.return_col_j(k).